[Feat] #162 - 복용 확인 후 알림 꺼도 약이 삭제되지 않도록 수정

+++ APillLog/APillLog/CoreData/CoreDataManager.swift
@@ -68,7 +68,7 @@ class CoreDataManager {
             
             for item in pillArray {
                 if (item.selectDate == todayDate && item.name == pill.name &&
-                    item.dosage == pill.dosage && ((item.cycle & pill.dosingCycle) != 0))
+                    item.dosage == pill.dosage && ((item.cycle & pill.dosingCycle) != 0) && !item.isTaking)
                 {
                     
                     self.context.delete(item)
